+++ src/ResourceManager/Sql/Commands.Sql.Test/ScenarioTests/ManagedDatabaseCrudScenarioTests.ps1
@@ -173,7 +173,9 @@ function Test-RestoreManagedDatabase
 {
 	# Setup
 	$rg = Create-ResourceGroupForTest
+	$rg2 = Create-ResourceGroupForTest
 	$managedInstance = Create-ManagedInstanceForTest $rg
+	$managedInstance2 = Create-ManagedInstanceForTest $rg2
 
 	try
 	{
@@ -187,16 +189,24 @@ function Test-RestoreManagedDatabase
 		Assert-AreEqual $db.Name $managedDatabaseName
 
 		$targetManagedDatabaseName = Get-ManagedDatabaseName
-		$pointInTime = (Get-date).AddMinutes(8)
+		$pointInTime = (Get-date).AddMinutes(5)
 
-		# Wait for 5 seconds for restore to be ready
-		Wait-Seconds 5
+		# Wait for 450 seconds for restore to be ready
+		Wait-Seconds 450
 
+		# restore managed database to the same instance
 		$restoredDb = Restore-AzureRmSqlManagedDatabase -FromPointInTimeBackup -ResourceGroupName $rg.ResourceGroupName -ManagedInstanceName $managedInstance.ManagedInstanceName -Name $managedDatabaseName -PointInTime $pointInTime -TargetManagedDatabaseName $targetManagedDatabaseName
 		Assert-NotNull $restoredDb
 		Assert-AreEqual $restoredDb.Name $targetManagedDatabaseName
 		Assert-AreEqual $restoredDb.ResourceGroupName $rg.ResourceGroupName
 		Assert-AreEqual $restoredDb.ManagedInstanceName $managedInstance.ManagedInstanceName
+
+		# restore managed database to the another instance, different resource group and managed instance
+		$restoredDb2 = Restore-AzureRmSqlManagedDatabase -FromPointInTimeBackup -ResourceGroupName $rg.ResourceGroupName -ManagedInstanceName $managedInstance.ManagedInstanceName -Name $managedDatabaseName -PointInTime $pointInTime -TargetManagedDatabaseName $targetManagedDatabaseName -TargetManagedInstanceName $managedInstance2.ManagedInstanceName -TargetResourceGroupName $rg2.ResourceGroupName
+		Assert-NotNull $restoredDb2
+		Assert-AreEqual $restoredDb2.Name $targetManagedDatabaseName
+		Assert-AreEqual $restoredDb2.ResourceGroupName $rg2.ResourceGroupName
+		Assert-AreEqual $restoredDb2.ManagedInstanceName $managedInstance2.ManagedInstanceName
 	}
 	finally
 	{